+++ gulpfile.ts
@@ -65,9 +65,9 @@ export class Gulpfile {
     @MergedTask()
     packageCompile() {
         const tsProject = ts.createProject("tsconfig.json");
-        const tsResult = gulp.src(["./src/**/*.ts", "./typings/**/*.ts"])
+        const tsResult = gulp.src(["./src/**/*.ts"])
             .pipe(sourcemaps.init())
-            .pipe(ts(tsProject));
+            .pipe(tsProject());
 
         return [
             tsResult.dts.pipe(gulp.dest("./build/package")),
@@ -118,15 +118,6 @@ export class Gulpfile {
     }
 
     /**
-     * This task will copy typings.json file to the build package.
-     */
-    @Task()
-    copyTypingsFile() {
-        return gulp.src("./typings.json")
-            .pipe(gulp.dest("./build/package"));
-    }
-
-    /**
      * Creates a package that can be published to npm.
      */
     @SequenceTask()
@@ -136,7 +127,7 @@ export class Gulpfile {
             "packageCompile",
             "packageMoveCompiledFiles",
             "packageClearCompileDirectory",
-            ["packagePreparePackageFile", "packageReadmeFile", "copyTypingsFile"]
+            ["packagePreparePackageFile", "packageReadmeFile"]
         ];
     }
 
@@ -174,7 +165,7 @@ export class Gulpfile {
         chai.should();
         chai.use(require("sinon-chai"));
         chai.use(require("chai-as-promised"));
-        return gulp.src("./build/es5/test/**/*.js")
+        return gulp.src("./build/compiled/test/**/*.js")
             .pipe(mocha());
     }
 
@@ -183,7 +174,7 @@ export class Gulpfile {
      */
     @Task()
     coveragePre() {
-        return gulp.src(["./build/es5/src/**/*.js"])
+        return gulp.src(["./build/compiled/src/**/*.js"])
             .pipe(istanbul())
             .pipe(istanbul.hookRequire());
     }
@@ -197,7 +188,7 @@ export class Gulpfile {
         chai.use(require("sinon-chai"));
         chai.use(require("chai-as-promised"));
         
-        return gulp.src(["./build/es5/test/**/*.js"])
+        return gulp.src(["./build/compiled/test/**/*.js"])
             .pipe(mocha())
             .pipe(istanbul.writeReports());
     }